move css near files
