+++ README.md
@@ -34,6 +34,4 @@ The SolidOS team is using GitHub Copilot integrated in Visual Studio Code.
 We have added comments in the code to make it explicit which parts are 100% written by AI.
 Example:
 * Some code was generated by the GPT-5.3-Codex model in GitHub Copilot based on the following prompt:
-* can you style my table and make it look nice.
-* can you handle the visibility in exposeOverlay with css?
-* move details section to be before the newTrackerButton
+* can you style the table and make it look nice.
