--- src/styles/utilities.css
@@ -1,4 +1,0 @@
-:root {
-  --icon-xxs: 1em;
-  --icon-xs: 1.5em;
-}
